+++ src/components/Event.js
@@ -3,7 +3,6 @@ import React from 'react'
 const Event = ({dispatch, event}) => {
   const id = event.id
   const handleClickDeleteButton = () => {
-    const result = window.confirm(`イベント(id=${id})を本当に削除してもいいですか？`)
     dispatch({type: 'DELETE_EVENT', id})
   }
